refactor: DOMUtils.js의 위치 변경 views -> utils
