+++ src/controllers/ProductAddManager.js
@@ -1,4 +1,4 @@
-import { $, default as DOM } from '../views/DOMUtils.js';
+import { $, default as DOM } from '../utils/DOMUtils.js';
 import { default as V } from '../utils/validators.js';
 import { default as DB } from '../model/database.js';
 import getAllPurchaseButton from './utils/getAllPurchaseButton.js';
