+++ src/controllers/VendingMachineChargeManager.js
@@ -1,4 +1,4 @@
-import { $, default as DOM } from '../views/DOMUtils.js';
+import { $, default as DOM } from '../utils/DOMUtils.js';
 import { default as V } from '../utils/validators.js';
 import convertChargeIntoCoin from './utils/convertChargeIntoCoin.js';
